+++ src/main/java/me/towdium/jecalculation/command/Commands.java
@@ -1,10 +1,10 @@
 package me.towdium.jecalculation.command;
 
 import mcp.MethodsReturnNonnullByDefault;
+import me.towdium.jecalculation.command.commands.CCraft;
 import me.towdium.jecalculation.command.commands.CHelp;
-import me.towdium.jecalculation.command.commands.COreDict;
+import me.towdium.jecalculation.command.commands.CMath;
 import me.towdium.jecalculation.command.commands.CState;
-import me.towdium.jecalculation.command.commands.CUuid;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 import java.util.HashMap;
@@ -21,8 +21,8 @@ public class Commands {
     static {
         add(new CHelp());
         add(new CState());
-        add(new COreDict());
-        add(new CUuid());
+        add(new CCraft());
+        add(new CMath());
     }
 
     static void add(ISubCommand c) {