+++ src/main/java/me/towdium/jecalculation/network/packets/PRecord.java
@@ -2,7 +2,7 @@ package me.towdium.jecalculation.network.packets;
 
 import io.netty.buffer.ByteBuf;
 import me.towdium.jecalculation.data.Controller;
-import me.towdium.jecalculation.data.structure.Recipes;
+import me.towdium.jecalculation.data.structure.RecordPlayer;
 import net.minecraft.nbt.NBTTagCompound;
 import net.minecraftforge.fml.common.network.ByteBufUtils;
 import net.minecraftforge.fml.common.network.simpleimpl.IMessage;
@@ -12,11 +12,12 @@ import net.minecraftforge.fml.common.network.simpleimpl.MessageContext;
 import java.util.Objects;
 
 public class PRecord implements IMessage {
-    public static final String KEY = "content";
-    NBTTagCompound recipes;
+    public static final String KEY_RECIPES = "recipes";
+    public static final String KEY_LAST = "last";
+    RecordPlayer record;
 
-    public PRecord(Recipes recipes) {
-        this.recipes = recipes.serialize();
+    public PRecord(RecordPlayer record) {
+        this.record = record;
     }
 
     public PRecord() {
@@ -25,20 +26,18 @@ public class PRecord implements IMessage {
     @Override
     public void fromBytes(ByteBuf buf) {
         NBTTagCompound tag = Objects.requireNonNull(ByteBufUtils.readTag(buf));
-        recipes = tag.getCompoundTag(KEY);
+        record = new RecordPlayer(tag);
     }
 
     @Override
     public void toBytes(ByteBuf buf) {
-        NBTTagCompound tag = new NBTTagCompound();
-        tag.setTag(KEY, recipes);
-        ByteBufUtils.writeTag(buf, tag);
+        ByteBufUtils.writeTag(buf, record.serialize());
     }
 
     public static class Handler implements IMessageHandler<PRecord, IMessage> {
         @Override
         public IMessage onMessage(PRecord message, MessageContext ctx) {
-            Controller.setRecipesServer(new Recipes(message.recipes));
+            Controller.setRecordsServer(message.record);
             return null;
         }
     }